all: revamp to parse csv from watsonbox/exportify
Rework everything to parse CSV outputs from watsonbox/exportify [1].
This is to provide some backlog data to ListenBrainz because it seems to
have some issues fetching history past a few months.

This also adds stricter type checking using mypy. This is my first time
attempting to dictate types in this fashion. I may have to use it more
in the future as it's good at pointing out when types are getting too
complex to be described reasonably in terms of only primitives.

[1] https://watsonbox.github.io/exportify/

Signed-off-by: Randolph Sapp <res.sapp@gmail.com>